Move `Product` and `ProductList` to `/react/Product`
